+++ packages/artrelay_core/lib/artrelay_core.dart
@@ -8,5 +8,3 @@ export 'src/media.dart';
 export 'src/models.dart';
 export 'src/pagination.dart';
 export 'src/repositories.dart';
-
-// TODO: Export any libraries intended for clients of this package.